Rename variable with inappropriate names
